Initial work with external script
