+++ Arduino/Drv/HardwareRateDriver/CMakeLists.txt
@@ -16,15 +16,15 @@ starts_with(IS_MBED "${ARDUINO_FQBN}" "arduino:mbed")
 if (NOT FPRIME_ARDUINO)
     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/HardwareRateDriverLinux.cpp")
 elseif (IS_TEENSY)
-    include_arduino_libraries("IntervalTimer")
+    target_use_arduino_libraries("IntervalTimer")
     list(APPEND MOD_DEPS "IntervalTimer")
     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/HardwareRateDriverTeensy.cpp")
 elseif (IS_MBED)
-    include_arduino_libraries("mbed")
+    target_use_arduino_libraries("mbed")
     list(APPEND MOD_DEPS "mbed")
     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/HardwareRateDriverMbed.cpp")
 else()
-    include_arduino_libraries("TimerOne")
+    target_use_arduino_libraries("TimerOne")
     list(APPEND MOD_DEPS "TimerOne")
     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/HardwareRateDriverAvr.cpp")
 endif()
